requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
cryptography>=41.0.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
loguru>=0.7.0
sentence-transformers>=2.2.2
torch  # Puedes fijar versión si lo deseas
numpy>=1.24.0
scikit-learn>=1.2.0
openai>=1.0.0
flake8>=6.1.0
bandit>=1.7.5
huggingface_hub[hf_xet]
gspread>=5.11.0
google-auth>=2.22.0
python-dateutil>=2.8.2
feedparser==6.0.11
beautifulsoup4==4.12.2
supabase>=2.0.0
psycopg2-binary>=2.9.9
httpx[http2]==0.27.*